Split table into submodules
